Add -Wall -Wextra, clean up warnings, migrate tar to stc::Unix::Process
Did run into a weird fucking bug with curl here. Haven't bothered doing
a clean build to rule out a fucked intermediate state as the cause, I
don't mind rebuilding curl. Weird fucking segfault though. I think
that's the first time I've run into gdb just giving up on the
stacktrace.

Also cleaned up some warnings in nodejs.lua

+++ src/upm/Context.cpp
@@ -113,7 +113,7 @@ void Context::parseFlags() {
                     }
                 } else {
                     // -f
-                    for (int i = 1; i < line.size(); ++i) {
+                    for (size_t i = 1; i < line.size(); ++i) {
                         if (shorthands.find(line[i]) != shorthands.end()) {
                             flags[shorthands.at(line[i])] = "true";
                         }
